moving some files around for consistancy
