--- hooks/api/types.ts
@@ -1,7 +1,0 @@
-export const baseQueryKey = 'passport-status:api:'
-
-/**
- * Check Status Endpoint
- */
-export const checkStatusQueryKey = `${baseQueryKey}:check-status`
-export const checkStatusUri = `/api/check-status`
