+++ pages/status.tsx
@@ -6,7 +6,7 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { useFormik } from 'formik'
 import * as Yup from 'yup'
 import { CheckStatusRequestBody } from './api/check-status'
-import { useCheckStatus } from '../hooks/api/useCheckStatus'
+import { useCheckStatus } from '../lib/CheckStatusHook'
 import Layout from '../components/Layout'
 import InputField from '../components/InputField'
 import ActionButton from '../components/ActionButton'
